% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_print-paste.R
\name{add_linebreaks}
\alias{add_linebreaks}
\title{Add line breaks to a linear predictor string}
\usage{
add_linebreaks(string, indent, width = 90L)
}
\arguments{
\item{string}{a character string (linear predictor)}

\item{indent}{integer; number of characters the new line should be indented}

\item{width}{integer; the maximum number of characters per line}
}
\description{
Adds line breaks to a string, breaking it after a "+" sign to not exceed a
given width of characters and taking into account indentation.
}
\keyword{internal}
