% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{all_vars}
\alias{all_vars}
\title{Extract names of variables from several objects}
\usage{
all_vars(...)
}
\arguments{
\item{...}{\code{formula} objects, lists of formulas or character strings that are
valid variable names (in the sense of \code{make.vars()})}
}
\description{
Version of \code{all.vars()} that can handle \code{formula}s, \code{lists} of \code{formulas} and
character strings.
}
\keyword{internal}
