% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{check_duplicate_groupings}
\alias{check_duplicate_groupings}
\title{Check for duplicate grouping levels}
\usage{
check_duplicate_groupings(groups)
}
\arguments{
\item{groups}{a list of integer "id" variables}
}
\value{
NULL; throws an error if duplicate grouping levels are found
}
\description{
Used within \code{get_groups()}.
}
\keyword{internal}
