% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{check_fixed_random}
\alias{check_fixed_random}
\title{Check wheather fixed or formula contains a random effects specification}
\usage{
check_fixed_random(arglist)
}
\arguments{
\item{arglist}{A list containing 'fixed', 'random', and 'formula' elements.}
}
\value{
The updated arglist.arglist
}
\description{
Checks if the objects provided to the \code{formula} and \code{fixed} arguments contain
a random effects specification. This function is used in random effects
models.
In case the combined fixed and random effects formula is part of the \code{fixed}
element, it is moved into the \code{formula} element.
}
\keyword{internal}
