% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_vcov.R
\name{check_full_blockdiag}
\alias{check_full_blockdiag}
\title{Replace a full with a block-diagonal variance covariance matrix
Check if a full random effects variance covariance matrix is specified
for a single variable. In that case, it is identical to a block-diagonal
matrix. Change the \code{rd_vcov} specification to \code{blockdiag} for clarity
(because then the variable name is used in the name of \code{b}, \code{D}, \code{invD}, ...)}
\usage{
check_full_blockdiag(rd_vcov)
}
\arguments{
\item{rd_vcov}{a valid random effects variance-covariance structure
specification (i.e., checked using \code{expand_rd_vcov_full()})}
}
\value{
a valid random effects variance-covariance structure specification
}
\description{
Replace a full with a block-diagonal variance covariance matrix
Check if a full random effects variance covariance matrix is specified
for a single variable. In that case, it is identical to a block-diagonal
matrix. Change the \code{rd_vcov} specification to \code{blockdiag} for clarity
(because then the variable name is used in the name of \code{b}, \code{D}, \code{invD}, ...)
}
\keyword{internal}
