% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{check_groups_vary_within_lvl}
\alias{check_groups_vary_within_lvl}
\title{Check if a grouping variable varies within another grouping variable}
\usage{
check_groups_vary_within_lvl(lvl, grouping_df)
}
\arguments{
\item{lvl}{a vector defining a grouping level}

\item{grouping_df}{a \code{data.frame} with where each column contains the
grouping indices of a hierarchical level (as obtained
by \code{get_groups(...)})}
}
\value{
a named logical vector indicating for each grouping variable
(column of \code{grouping_df}) whether it varies within the grouping
defined by \code{lvl}
}
\description{
Check if a grouping variable varies within another grouping variable
}
\keyword{internal}
