% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{check_na_groupings}
\alias{check_na_groupings}
\title{Check for missing values in grouping variables}
\usage{
check_na_groupings(groups)
}
\arguments{
\item{groups}{a list of integer "id" variables}
}
\value{
\code{NULL}; throws an error if missing values are found in any grouping
variable
}
\description{
Checks if any grouping variable contains missing values (\code{NA} or \code{NaN}) and
throws an error if so. Used within \code{get_groups()}.
}
\keyword{internal}
