% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_divide_matrices.R
\name{check_rd_vcov}
\alias{check_rd_vcov}
\title{Check / create the random effects variance-covariance matrix specification}
\usage{
check_rd_vcov(rd_vcov, nranef)
}
\arguments{
\item{rd_vcov}{variance covariance specification provided by the user}

\item{nranef}{list by level with named vectors of number of random effects
per variable (obtained by \code{get_nranef()})}
}
\description{
Check / create the random effects variance-covariance matrix specification
}
\keyword{internal}
