% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_divide_matrices.R
\name{check_rd_vcov_list}
\alias{check_rd_vcov_list}
\title{First validation for rd_vcov}
\usage{
check_rd_vcov_list(rd_vcov, idvar)
}
\arguments{
\item{rd_vcov}{the random effects variance covariance structure provided by
the user}

\item{idvar}{vector with the names of the grouping variables
(without "lvlone")}
}
\description{
Check if rd_vcov is a list with elements for all grouping levels or does
not specify a grouping level. If valid, make sure it is a list per grouping
level by duplicating the contents if necessary.
}
