% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{check_redundant_lvls}
\alias{check_redundant_lvls}
\title{Check for unnecessary grouping levels}
\usage{
check_redundant_lvls(groups, nrow_data)
}
\arguments{
\item{groups}{a list of integer "id" variables}
}
\value{
NULL; throws an error if unnecessary grouping levels are found
}
\description{
Used within \code{get_groups()}.
}
\keyword{internal}
