% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{check_tvar}
\alias{check_tvar}
\title{Check if a variable is time-varying}
\usage{
check_tvar(x, idvar)
}
\arguments{
\item{x}{a vector, the variable to be tested}

\item{idvar}{a vector specifying a grouping}
}
\value{
a logical value
}
\description{
Check if a variable is time-varying
}

