% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{check_unnecessary_grouping_levels}
\alias{check_unnecessary_grouping_levels}
\title{Check for unnecessary grouping levels}
\usage{
check_unnecessary_grouping_levels(groups, nrow_data)
}
\arguments{
\item{groups}{a list of grouping information, as obtained from \code{get_groups()}}
}
\value{
NULL; throws an error if unnecessary grouping levels are found
}
\description{
Check for unnecessary grouping levels
}
\keyword{internal}
