% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{check_vars_in_data}
\alias{check_vars_in_data}
\title{Check that all variables in formulas are in the data}
\usage{
check_vars_in_data(datanames, fixed = NULL, random = NULL,
  auxvars = NULL, timevar = NULL)
}
\arguments{
\item{datanames}{a character vector (of all variable names in the data)}

\item{fixed}{the fixed effects formula (or list of formulas)}

\item{random}{the random effects formula (or list of formulas)}

\item{auxvars}{one-sided formula of auxiliary variables}

\item{timevar}{a character string (name of the time variable, used in joint models)}
}
\value{
nothing, but throws an error if a variable is missing
}
\description{
Check that all variables in formulas are in the data
}
\keyword{(2020-06-09)}
\keyword{in}
\keyword{internal}
\keyword{model_imp}
\keyword{used}
