% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_divide_matrices.R
\name{choose_default_model}
\alias{choose_default_model}
\title{Choose default analysis model based on outcome and data level}
\usage{
choose_default_model(outcome, lvl, max_lvl)
}
\arguments{
\item{outcome}{outcome variable}

\item{lvl}{hierarchical level of the outcome variable}

\item{max_lvl}{maximum hierarchical level in the data}
}
\value{
character string giving the default model type
}
\description{
Choose default analysis model based on outcome and data level
}
\keyword{internal}
