% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{clean_names}
\alias{clean_names}
\title{Replace ":" with "_" in a string}
\usage{
clean_names(string)
}
\arguments{
\item{string}{a character string}
}
\value{
the cleaned character string
}
\description{
Cleans up factor levels (or other strings) by replacing ":" with "_" to avoid
issues with the current implementation of identifying interactions (which looks
for ":" in model terms).
}
\details{
used in this file (in convert_variables()) (2020-06-09)
}
\keyword{internal}
