% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{combine_formula_lists}
\alias{combine_formula_lists}
\title{Combine fixed and random effects formulas}
\usage{
combine_formula_lists(fixed, random, warn = TRUE)
}
\arguments{
\item{fixed}{a fixed effects formula or list of such formulas}

\item{random}{a random effects formula (only RHS) or list of such formulas}

\item{warn}{logical; should the warning(s) be printed}
}
\description{
A function to combine nlme-style fixed and random effects formulas into
lme4 style formulas.
}
\details{
Internal function.
Lists of formulas can be named or unnamed.
Uses \code{combine_formulas()}.
}
\keyword{internal}
