% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{combine_formulas}
\alias{combine_formulas}
\title{Combine a fixed and random effects formula}
\usage{
combine_formulas(fixed, random)
}
\arguments{
\item{fixed}{fixed effects formula (two-sided \code{formula} object)}

\item{random}{random effects formula (one-sided \code{formula} object)}
}
\description{
Combine a single fixed and random effects formula by pasting them together.
}
\details{
Internal function, used in \code{combine_formula_lists()}.
}
\keyword{internal}
