% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{compare_data_structure}
\alias{compare_data_structure}
\title{Compare the structure of two data.frames}
\usage{
compare_data_structure(data1, data2)
}
\arguments{
\item{data1}{a \code{data.frame}}

\item{data2}{a \code{data.frame}}
}
\value{
nothing, but prints messages if the class of any variable
changed or if the levels of any factor variable changed
}
\description{
Compare the structure of two data.frames
}
\keyword{internal}
