% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{densplot}
\alias{densplot}
\alias{densplot.JointAI}
\title{Densplots}
\usage{
densplot(object, ...)

\method{densplot}{JointAI}(object, start = NULL, end = NULL, thin = NULL,
  subset = "main", vlines = NULL, nrow = NULL, ncol = NULL, ...)
}
\arguments{
\item{object}{object}
}
\description{
Densplots
}
