% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGS.R
\name{difftime_df}
\alias{difftime_df}
\title{Converts a \code{difftime} object to a \code{data.frame}}
\usage{
difftime_df(dt)
}
\arguments{
\item{dt}{\code{difftime} object (vector of \code{difftime} objects)}
}
\description{
Converts a \code{difftime} object to a \code{data.frame}
}
\keyword{internal}
