% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_matrices.R
\name{divide_matrices}
\alias{divide_matrices}
\title{Create data matrices for time constant and time-varying variables}
\usage{
divide_matrices(DF, fixed, random = NULL, auxvars = NULL, center = T,
  scale = T)
}
\arguments{
\item{DF}{a dataframe}

\item{fixed}{a formula describing the mean structure}

\item{random}{an optional formula describing the random effects (check this!)}

\item{auxvars}{vector containing the names of auxiliary variables}

\item{center}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}

\item{scale}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}
}
\value{
a list containing the matrices
}
\description{
Create data matrices for time constant and time-varying variables
}

