% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{drop_levels}
\alias{drop_levels}
\title{Check for empty factor levels}
\usage{
drop_levels(data, allvars, warn = TRUE)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{allvars}{a character vector (of all variable names used in the model)}

\item{warn}{logical, if \code{TRUE} warnings are printed}
}
\value{
the \code{data.frame} (unchanged)
}
\description{
CHECKS if there are empty factor levels in any of the variables used in the
model.
}
\section{Note}{

Originally, the function also dropped these levels. Then, I (accidentally?)
had commented out a line so that no check was performed.
Now, only create a warning if there are empty levels, but do not drop them.

used in model_imp (2020-06-09)
}

\keyword{internal}
