% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGS.R
\name{duration_obj}
\alias{duration_obj}
\title{Create a duration object}
\usage{
duration_obj(dur)
}
\arguments{
\item{dur}{list of \code{difftime} objects}
}
\description{
Add row names to the object
}
\keyword{internal}
