% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_vcov.R
\name{expand_rd_vcov_full}
\alias{expand_rd_vcov_full}
\title{Expand rd_vcov using variable names in case "full" is used}
\usage{
expand_rd_vcov_full(rd_vcov, rd_outnam)
}
\arguments{
\item{rd_vcov}{the random effects variance covariance structure provided by
the user (\code{check_rd_vcov_list()} is called internally)}

\item{rd_outnam}{list by grouping level of the names of the outcome variables
that have random effects on this level}
}
\value{
A named list per grouping level where each elements contains
information on how the random effects variance-covariance matrices on
that level are structured. Per level there is a list of grouping
structures containing the names of variables in each structure
(e.g. \code{list(full = c("a", "b"), indep = "c")})
}
\description{
Expand rd_vcov using variable names in case "full" is used
}
\keyword{internal}
