% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{extract_fixef_formula}
\alias{extract_fixef_formula}
\title{Extract fixed effects formula from lme4-type formula}
\usage{
extract_fixef_formula(formula)
}
\arguments{
\item{formula}{a \code{formula} object (typically in lme4-style)}
}
\value{
a \code{formula} object
}
\description{
Extract fixed effects formula from lme4-type formula
}
\keyword{internal}
