% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{extract_grouping}
\alias{extract_grouping}
\title{Extract grouping variables from a (list of) formula(s)}
\usage{
extract_grouping(formula, warn = FALSE)
}
\arguments{
\item{formula}{a \code{formula} object or a \code{list} of \code{formula} objects}

\item{warn}{does nothing}
}
\value{
a vector of character strings containing the unique grouping
variable names found in any of the input formulas, or \code{NULL} if
there is no grouping found
}
\description{
Extracts the grouping variables (i.e., variables after \code{|}) from a model
formula or a list of such formulas
}
\keyword{internal}
