% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{extract_lhs_string}
\alias{extract_lhs_string}
\title{Extract the left hand side of a formula}
\usage{
extract_lhs_string(formula)
}
\arguments{
\item{formula}{a \code{formula} object (NOT a \code{list} of formulas)}
}
\value{
A character string.
}
\description{
Extracts the left hand side from a \code{formula} object and returns it as
character string.
}
\details{
Internal; used in various help functions (2022-02-05)
}
\keyword{internal}
