% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{extract_ranef_formula}
\alias{extract_ranef_formula}
\title{Extract random effects formula from lme4-type formula}
\usage{
extract_ranef_formula(formula)
}
\arguments{
\item{formula}{a \code{formula} object (typically in lme4-style)}
}
\value{
a one-sided \code{formula} object (or \code{NULL} if there are no random
effects)
}
\description{
Extract random effects formula from lme4-type formula
}
\keyword{internal}
