% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_state.R
\name{extract_state}
\alias{extract_state}
\title{Return the current state of a 'JointAI' model}
\usage{
extract_state(object, pattern = paste0("^", c("RinvD", "invD", "tau", "b"),
  "_"))
}
\arguments{
\item{object}{an object of class 'JointAI'}

\item{pattern}{vector of patterns to be matched with the names of the nodes}
}
\value{
A list with one element per chain of the MCMC sampler, containing the
Returns the current state of the MCMC sampler (values of the last
iteration) for the subset of nodes identified based on the pattern
the user has specified.
}
\description{
Return the current state of a 'JointAI' model
}
