% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{extract_y}
\alias{extract_y}
\title{Extract outcome variable from fixed effects formula}
\usage{
extract_y(fixed)
}
\arguments{
\item{fixed}{a formula describing fixed effects structure}
}
\value{
a character string, giving the name of the outcome variable
}
\description{
Extract outcome variable from fixed effects formula
}
