% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_divide_matrices.R
\name{factor_to_integer}
\alias{factor_to_integer}
\title{Convert a factor to an integer representation}
\usage{
factor_to_integer(x)
}
\arguments{
\item{x}{a vector. If \code{x} is a factor it will be converted to integers as
described below; otherwise \code{x} is returned unchanged.}
}
\value{
An integer vector when \code{x} is a factor, otherwise the original \code{x}.
}
\description{
Convert a factor to an integer representation
}
\note{
Used in \code{extract_outcome_data()}.
}
\keyword{internal}
