% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{find_continuous}
\alias{find_continuous}
\title{Function to find the names of columns in the model matrix that involve
continuous covariates (and hence may need to be scaled)
for now not used}
\usage{
find_continuous(fixed, DF, contr = NULL)
}
\description{
Function to find the names of columns in the model matrix that involve
continuous covariates (and hence may need to be scaled)
for now not used
}
