% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{find_continuous_main}
\alias{find_continuous_main}
\title{Function to find the names of columns in the model matrix that involve
continuous covariates (and hence may need to be scaled) - only main effects}
\usage{
find_continuous_main(fixed, DF)
}
\description{
Function to find the names of columns in the model matrix that involve
continuous covariates (and hence may need to be scaled) - only main effects
}
