% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{find_positions}
\alias{find_positions}
\title{Find position of variable names in a vector of variable names}
\usage{
find_positions(nams1, nams2)
}
\arguments{
\item{nams1}{vector of variable name to look for}

\item{nams2}{vector of variable names}
}
\value{
integer
}
\description{
Find position of variable names in a vector of variable names
}
