% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{gen_pat}
\alias{gen_pat}
\title{Generate pattern (used in get_hc_list)}
\usage{
gen_pat(nam)
}
\arguments{
\item{nam}{variable name (character string)}
}
\value{
vector
}
\description{
Generate pattern (used in get_hc_list)
}
