% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_names.R
\name{get_coef_names}
\alias{get_coef_names}
\title{extract names of regression coefficients}
\usage{
get_coef_names(Mlist, K)
}
\arguments{
\item{Mlist}{Mlist}

\item{K}{K}
}
\description{
extract names of regression coefficients
}
