% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_list.R
\name{get_data_list}
\alias{get_data_list}
\title{Create list of data passed to JAGS}
\usage{
get_data_list(analysis_type, family, meth, Mlist, K, auxvars,
  scale_pars = NULL)
}
\arguments{
\item{analysis_type}{analysis model type}

\item{meth}{vector of imputation methods}

\item{Mlist}{list of data matrices etc.}
}
\description{
Create list of data passed to JAGS
}
