% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{get_dest_column}
\alias{get_dest_column}
\title{Find which column in either Xc or Xcat contains the variable to be imputed}
\usage{
get_dest_column(varname, DF, Xc_names, Xcat_names)
}
\arguments{
\item{varname}{variable name}

\item{Xc_names}{column names of the design matrix of baseline effects}

\item{Xcat_names}{column names of the matrix of categorical variables}

\item{impmeth}{imputation method}
}
\description{
Find which column in either Xc or Xcat contains the variable to be imputed
}

