% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_info.R
\name{get_family}
\alias{get_family}
\title{Identify the family from the covariate model type}
\usage{
get_family(model)
}
\arguments{
\item{model}{character string; the covariate model type}
}
\description{
Identify the family from the covariate model type
}
\keyword{internal}
