% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{get_grouping_levels}
\alias{get_grouping_levels}
\title{Get grouping levels}
\usage{
get_grouping_levels(grouping_df)
}
\arguments{
\item{grouping_df}{a \code{data.frame} of grouping ("id") variables, as obtained
from \code{get_groups()}}
}
\value{
a named integer vector identifying the grouping hierarchy, where
"1" is the level of the individual observations and the largest
number is the highest grouping level. Crossed levels will have the
same number.
}
\description{
A helper function that identifies the hierarchy of grouping variables. It
checks for each grouping level how many of the other grouping levels it
varies within and returns the ranked order of the grouping levels.
}
\keyword{internal}
