% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{get_groups}
\alias{get_groups}
\title{Check if a variable is time-varying}
\usage{
get_groups(idvar, data)
}
\arguments{
\item{idvar}{a vector specifying a grouping}

\item{x}{a vector, the variable to be tested}
}
\value{
a logical value
}
\description{
Check if a variable is time-varying
}
\keyword{internal}
