% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{get_hc_list}
\alias{get_hc_list}
\title{Hierarchical centering structure}
\usage{
get_hc_list(X2, Xc, Xic, Z, Xlong)
}
\arguments{
\item{X2_names}{vector of column names in the extended fixed effects design matrix}

\item{Z_names}{vector of column names of the random effects design matrix}

\item{Xc_names}{vector of column names of the matrix of baseline covariates}
}
\value{
A named list with an entry for each level of random effects, omitting
        the random intercept. Each list entry is a named vector, where the
        names represent the variable names of either the main effect of a
        variable that is part of the random and fixed effects structure, or
        the name of a variable that has an interaction with the variable for
        which there is a random effect. The values represent the column
        number of the respective matrix for the variable (for interaction terms
        the column number of the variable that is not the respective
        random effect). The attribute
        "matrix" names the matrix (Z or Xc), and is NA when the variable
        is not in either of the two matrices.
}
\description{
Hierarchical centering structure
}
