% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_dimensions.R
\name{get_imp_dim}
\alias{get_imp_dim}
\title{Determine number of parameters in the imputation models}
\usage{
get_imp_dim(meth, pos_Xc)
}
\arguments{
\item{meth}{named vector specifying the imputation methods and ordering of
the imputation models}

\item{pos_Xc}{a list containing the positions of the incomplete variables in Xc}
}
\value{
a matrix specifying the range of regression coefficients per
        imputation model
}
\description{
Determine number of parameters in the imputation models
}
