% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imp_method.R
\name{get_imp_meth}
\alias{get_imp_meth}
\title{Set default imputation methods}
\usage{
get_imp_meth(DF, fixed = NULL, random = NULL, id = NULL, auxvars = NULL)
}
\arguments{
\item{DF}{a dataframe}

\item{fixed}{a formula}

\item{random}{a formula specifying a random effects structure}

\item{id}{a character string identifying the id variable if there is a
grouping, otherwise NULL}

\item{auxvars}{vector of variable names as auxiliary variables}
}
\value{
a named vector containing those variables in DF
        (check what happens if DF doesn't match fixed) that have missing
        values and their assigned default imputation methods, sorted by
        proportion of missing values (first cross-sectional variables,
        then longitudinal variables)
}
\description{
Set default imputation methods
}

