% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{get_imp_par_list}
\alias{get_imp_par_list}
\title{Creates a list of parameters that will be passed to the functions generating the imputation models}
\usage{
get_imp_par_list(impmeth, varname, Xc, Xcat, K_imp, dest_cols)
}
\arguments{
\item{impmeth}{character string specifying the imputation method}

\item{varname}{name of the variable to be imputed}

\item{Xc}{design matrix of time-constant covariates}

\item{Xcat}{matrix of incomplete categorical covariates with more than 2 categories}

\item{K_imp}{matrix specifying the range of elements of the vector of regression
coefficients to be used in each imputation model}

\item{imp_pos}{list specifying which column(s) of Xc contain(s) the variable
to be imputed}
}
\description{
Creates a list of parameters that will be passed to the functions generating the imputation models
}

