% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_dimensions.R
\name{get_imp_pos}
\alias{get_imp_pos}
\title{Determine positions of incomplete variables in the data matrices}
\usage{
get_imp_pos(meth, DF, Mlist)
}
\arguments{
\item{meth}{named vector specifying the imputation methods and ordering of
the imputation models}

\item{DF}{data frame}

\item{Mlist}{a named list with the entries "Xc", "Xic", "Xl", "Xil", "Z"}
}
\value{
a list?
}
\description{
Determine positions of incomplete variables in the data matrices
}

