% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_vcov.R
\name{get_listelement}
\alias{get_listelement}
\title{Get an element of a list, return a default value if it does not exist}
\usage{
get_listelement(object, element, null_value = 0)
}
\arguments{
\item{object}{a \code{list}}

\item{element}{the name of the element to extract (a character string)}

\item{null_value}{the value to return if the element does not exist}
}
\value{
the value of \code{object[[element]]} or \code{null_value} if
\code{object[[element]]} is \code{NULL}
}
\description{
A small helper function to extract an element of a list and return a default
value if the element does not exist (i.e., is \code{NULL}).
}
\examples{
mylist <- list(A = list(a = 1, b = 2), B = list(a = 23))
sapply(mylist, get_listelement, "b", null_value = 0)
# returns c(A = 2, B = 0)
}
\keyword{internal}
