% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{get_mis_meth}
\alias{get_mis_meth}
\title{Set default imputation methods}
\usage{
get_mis_meth(DF, fixed, id = NULL)
}
\arguments{
\item{DF}{a dataframe}

\item{fixed}{a formula}

\item{id}{a character string identifying the id variable if there is a
grouping, otherwise NULL}
}
\value{
a named vector containing those variables in DF
        (check what happens if DF doesn't match fixed) that have missing
        values and their assigned default imputation methods, sorted by
        proportion of missing values (first cross-sectional variables,
        then longitudinal variables)
}
\description{
Set default imputation methods
}

