% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.JointAI.R
\name{get_missinfo}
\alias{get_missinfo}
\title{Obtain a summary of the missing values involved in an object of class JointAI}
\usage{
get_missinfo(object)
}
\arguments{
\item{object}{object inheriting from class JointAI}
}
\description{
This function returns a \code{data.frame} or a \code{list} of
\code{data.frame}s per grouping level. Each of the \code{data.frames}
has columns \code{variable}, \code{#NA} (number of missing values) and
\code{\%NA} (proportion of missing values in percent).
}
\examples{
mod <-  lm_imp(y ~ C1 + B2 + C2, data = wideDF, n.iter = 100)
get_missinfo(mod)


}
