% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_matrices.R
\name{get_Mlist}
\alias{get_Mlist}
\title{Re-create the full \code{Mlist} from a "JointAI" object}
\usage{
get_Mlist(object)
}
\arguments{
\item{object}{object of class "JointAI"}
}
\description{
Re-create the full \code{Mlist} from a "JointAI" object
}
\keyword{internal}
