% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_dimensions.R
\name{get_model_dim}
\alias{get_model_dim}
\title{Determine number of fixed effects / regression coefficients in the analysis model}
\usage{
get_model_dim(ncols, hc_list)
}
\arguments{
\item{ncols, }{named list specifying the column numbers of the matrices Xc,
Xic, Xl and Xil}

\item{hc_list}{named vector or list specifying hierarchical centering
structure (see \code{\link{get_hc_list}})}
}
\value{
a matrix specifying the range of regression coefficients per
        component of the analysis model
}
\description{
Determine number of fixed effects / regression coefficients in the analysis model
}
\note{
Auxiliary variables are treated the same way as variables that are
      actually in the model.
}

