% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_info.R
\name{get_modeltype}
\alias{get_modeltype}
\title{Identify the general model type from the covariate model type}
\usage{
get_modeltype(model)
}
\arguments{
\item{model}{character string; the covariate model type}
}
\description{
Identify the general model type from the covariate model type
}
\keyword{internal}
