% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_vcov.R
\name{get_nranef}
\alias{get_nranef}
\title{Extract the number of random effects}
\usage{
get_nranef(random, data)
}
\arguments{
\item{random}{a random effect formula or list of random effects formulas}

\item{data}{a \code{data.frame}}

\item{idvar}{vector of the names of all id variables}
}
\value{
a list by grouping level (\code{idvar}) with a named vector of the number
of random effects per variable (=names).
}
\description{
Extract the number of random effects
}
\keyword{internal}
