% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get parameters to follow}
\usage{
get_params(meth, analysis_type, Xc, Xcat, y_name = NULL, Zcols = NULL,
  analysis_main = F, analysis_random = F, imputation_pars = F,
  imp_pars = F, imps = NULL, betas = NULL, tau_y = NULL,
  sigma_y = NULL, ranef = NULL, invD = NULL, D = NULL, RinvD = NULL,
  alphas = NULL, tau_imp = NULL, gamma_imp = NULL, delta_imp = NULL,
  other = NULL)
}
\arguments{
\item{meth}{imputation method}

\item{analysis_type}{analysis model type}

\item{Xc}{matrix}

\item{Xcat}{matrix}

\item{y_name}{name of the outcome variable}

\item{Zcols}{number of columns in random effects design matrix}

\item{analysis_main}{logical}

\item{analysis_random}{logical}

\item{imputation_pars}{logical}

\item{imp_pars}{logical}

\item{imps}{logical}

\item{betas}{logical}

\item{tau_y}{logical}

\item{sigma_y}{logical}

\item{ranef}{logical}

\item{invD}{logical}

\item{D}{logical}

\item{RinvD}{logical}

\item{alphas}{logical}

\item{tau_imp}{logical}

\item{gamma_imp}{logical}

\item{delta_imp}{logical}
}
\description{
Get parameters to follow
}

