% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{get_refcat}
\alias{get_refcat}
\title{Specify the reference category for a categorical variable}
\usage{
get_refcat(varname, Xcat, refcats)
}
\arguments{
\item{varname}{name of the variable}

\item{Xcat}{matrix of categorical covariates}

\item{refcats}{character string or named vector of reference categories for
each categorical variable}
}
\description{
Specify the reference category for a categorical variable
}
