% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_info_list.R
\name{get_resp_mat}
\alias{get_resp_mat}
\title{Identify the data matrix containing a given response variable}
\usage{
get_resp_mat(resp, Mlvls, outnames)
}
\arguments{
\item{resp}{character string; name of the response variable}

\item{Mlvls}{named vector where the names are all column names of all data
matrices, and the values are the names of the corresponding
data matrices}

\item{outnames}{character vector; names of the columns in the data matrices
that contain the response variable (or multiple columns in
case of a survival outcome)}
}
\value{
character string; the name(s) of the data matrix/matrices of the
response variable(s)
}
\description{
Identify the data matrix containing a given response variable
}
\keyword{internal}
