% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{get_X}
\alias{get_X}
\title{Create data matrices for time constant and time-varying variables}
\usage{
get_X(DF, fixed, random, id)
}
\arguments{
\item{DF}{a dataframe}

\item{fixed}{a formula describing the mean structure}

\item{random}{an optional formula describing the random effects (check this!)}

\item{id}{a character string identifying the id variable if there is a
grouping, otherwise NULL}
}
\value{
a list containing the matrices
}
\description{
Create data matrices for time constant and time-varying variables
}

