% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_glm.R
\name{glm_model}
\alias{glm_model}
\title{Function to write generalized linear regression model as analysis model}
\usage{
glm_model(family, link, N, y_name, Xic = NULL, K, ...)
}
\arguments{
\item{N}{number of subjects / random intercepts}

\item{Xic}{design matrix of cross-sectional interaction effects}

\item{K}{matrix specifying the number of parameters for each component of the
fixed effects}

\item{y}{y}

\item{Z}{random effects design matrix}

\item{Xl}{design matrix of longitudinal covariates}

\item{hc_list}{hierarchical centering specification}
}
\description{
Function to write generalized linear regression model as analysis model
}
