% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_glm.R
\name{glm_priors}
\alias{glm_priors}
\title{Write priors for the regression coefficients of the linear model}
\usage{
glm_priors(K, y_name, ...)
}
\arguments{
\item{K}{K}

\item{y_name}{character string, name of outcome}
}
\description{
Write priors for the regression coefficients of the linear model
}
