% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_criteria.R
\name{gr_crit}
\alias{gr_crit}
\title{Gelman-Rubin criterion for convergence}
\usage{
gr_crit(object, confidence = 0.95, transform = FALSE, autoburnin = TRUE,
  multivariate = TRUE, subset = "main", start = NULL, end = NULL,
  thin = NULL)
}
\arguments{
\item{object}{JointAI object}

\item{confidence}{the coverage probability of the confidence interval for the
potential scale reduction factor}

\item{transform}{a logical flag indicating whether variables in
\code{x} should be transformed to improve the normality of the
distribution. If set to TRUE, a log transform or logit transform, as
appropriate, will be applied.}

\item{autoburnin}{a logical flag indicating whether only the second half
  of the series should be used in the computation.  If set to TRUE
  (default) and \code{start(x)} is less than \code{end(x)/2} then start
  of series will be adjusted so that only second half of series is used.}

\item{multivariate}{a logical flag indicating whether the multivariate
potential scale reduction factor should be calculated for multivariate
chains}
}
\description{
Gelman-Rubin criterion for convergence
}
