% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{grep_names}
\alias{grep_names}
\title{Find names in a vector of names}
\usage{
grep_names(nams1, nams2)
}
\arguments{
\item{nams1}{the names to look for (generated by gen_pat)}

\item{nams2}{the vector of names to search in}
}
\description{
Find names in a vector of names
}
