% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_continuous.R
\name{impmodel_lognormal}
\alias{impmodel_lognormal}
\title{Imputation by Bayesian log-normal regression}
\usage{
impmodel_lognormal(varname, dest_col, dest_mat, Xc_cols, par_elmts, par_name,
  ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{dest_col}{column of Xc containing the variable to be imputed}

\item{Xc_cols}{columns of the design matrix to used in linear predictor}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Imputation by Bayesian log-normal regression
}
