% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_continuous.R
\name{impmodel_normal}
\alias{impmodel_normal}
\title{Imputation by Bayesian linear regression}
\usage{
impmodel_normal(varname, dest_col, dest_mat, trafo_cols, trafos, trfo_fct,
  Xc_cols, par_elmts, par_name, ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{dest_col}{column of Xc containing the variable to be imputed}

\item{Xc_cols}{columns of the design matrix to used in linear predictor}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Imputation by Bayesian linear regression
}
