% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_ordinal.R
\name{impmodel_ordinal}
\alias{impmodel_ordinal}
\title{Imputation by cumulative logistic regression}
\usage{
impmodel_ordinal(varname, dest_col, Xc_cols, par_elmts, par_name, dummy_cols,
  ncat, refcat, ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{dest_col}{column of Xc containing the variable to be imputed}

\item{Xc_cols}{columns of the design matrix to used in linear predictor}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Imputation by cumulative logistic regression
}

