% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_binary.R
\name{impprior_logit}
\alias{impprior_logit}
\title{Priors for logistic imputation model}
\usage{
impprior_logit(varname, par_elmts, par_name, ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Priors for logistic imputation model
}
