% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_multinomial.R
\name{impprior_multinomial}
\alias{impprior_multinomial}
\title{Priors for multinomial imputation model}
\usage{
impprior_multinomial(varname, par_elmts, par_name, ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Priors for multinomial imputation model
}

