% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_continuous.R
\name{impprior_normal}
\alias{impprior_normal}
\title{Priors for linear imputation model}
\usage{
impprior_normal(varname, par_elmts, par_name, ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Priors for linear imputation model
}

