% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmodels_ordinal.R
\name{impprior_ordinal}
\alias{impprior_ordinal}
\title{Priors for ordinal imputation model}
\usage{
impprior_ordinal(varname, par_elmts, par_name, ncat, ...)
}
\arguments{
\item{varname}{name of the variable to be imputed}

\item{par_elmts}{elements of the parameter vector to be used}

\item{par_name}{name of the parameter}
}
\description{
Priors for ordinal imputation model
}

