% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_imp.R
\name{lm_imp}
\alias{lm_imp}
\title{Main function for analysis with a linear model}
\usage{
lm_imp(fixed, data, auxvars = NULL, refcats = "largest", modelfile = NULL,
  n.chains = 3, n.adapt = 100, n.iter = 0, runJAGS = F)
}
\arguments{
\item{fixed}{a two sided formula describing the fixed-effects part of the
model (see \code{\link[stats]{formula}})}

\item{data}{a data frame containing the variabels named in \code{fixed}}

\item{auxvars}{an optional vector of variable names that shoud be included
as predictors in the imputation procedure but are not part of
the analysis model}

\item{refcats}{a named vector specifying which category should be treated as
as reference category for each categorical variable, "first"
or "largest" (default). When \code{refcat = "largest"}, the
category with the most observations is chosen for each
categorical variable.}

\item{modelfile}{optional name (and path) of the file the JAGS model will be
written to}

\item{n.chains}{number of mcmc chains}

\item{n.adapt}{number of iterations in adaptive phase}

\item{n.iter}{number of iterations in sampling}

\item{runJAGS}{logical}
}
\description{
Main function for analysis with a linear model
}

