% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_lm.R
\name{lm_priors}
\alias{lm_priors}
\title{Write priors for the regression coefficients of the linear model}
\usage{
lm_priors(K, y_name, ...)
}
\arguments{
\item{K}{K}

\item{y_name}{character string, name of outcome}
}
\description{
Write priors for the regression coefficients of the linear model
}

