% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_imp.R
\name{lme_imp}
\alias{lme_imp}
\title{Main function for analysis with a linear mixed model}
\usage{
lme_imp(fixed, data, random, auxvars = NULL, refcats = "largest",
  monitor_params = NULL, modelfile = NULL, n.chains = 3, n.adapt = 100,
  n.iter = 100, runJAGS = FALSE, center = T, scale = T)
}
\arguments{
\item{fixed}{a two sided formula describing the fixed-effects part of the
model (see \code{\link[stats]{formula}})}

\item{data}{a data frame containing the variabels named in \code{fixed} and
\code{random}}

\item{random}{a one sided formula of the form "\code{ ~ x | id}", where
\code{x} is the variable that has a random effect and \code{id}
is the variable specifying the grouping structure.}

\item{auxvars}{an optional vector of variable names that shoud be included
as predictors in the imputation procedure but are not part of
the analysis model}

\item{refcats}{a named vector specifying which category should be treated as
as reference category for each categorical variable, "first"
or "largest" (default). When \code{refcat = "largest"}, the
category with the most observations is chosen for each
categorical variable.}

\item{monitor_params}{a character vector giving the names of variables to be monitored, see details}

\item{modelfile}{optional name (and path) of the file the JAGS model will be
written to}

\item{n.chains}{the number of parallel chains for the model}

\item{n.adapt}{the number of iterations for adaptation. See 
    \code{\link{adapt}} for details. If \code{n.adapt = 0} then no 
    adaptation takes place.}

\item{n.iter}{number of iterations to monitor}

\item{runJAGS}{logical}

\item{center}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}

\item{scale}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}
}
\description{
Main function for analysis with a linear mixed model
}

