% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_lme.R
\name{lme_model}
\alias{lme_model}
\title{Function to write linear mixed model as analysis model}
\usage{
lme_model(N, y_name, Z = NULL, Xic = NULL, Xl = NULL, Xil = NULL,
  hc_list = NULL, Mlist = NULL, K, ...)
}
\arguments{
\item{N}{number of subjects / random intercepts}

\item{Z}{random effects design matrix}

\item{Xic}{design matrix of cross-sectional interaction effects}

\item{Xl}{design matrix of longitudinal covariates}

\item{hc_list}{hierarchical centering specification}

\item{K}{matrix specifying the number of parameters for each component of the
fixed effects}

\item{y}{y}
}
\description{
Function to write linear mixed model as analysis model
}
