% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_lme.R
\name{lme_priors}
\alias{lme_priors}
\title{Write priors for a linear mixed model}
\usage{
lme_priors(K, y_name, Z = NULL, Mlist = NULL, ...)
}
\arguments{
\item{K}{K}

\item{y_name}{name of the outcome}

\item{Z}{random effects design matrix}
}
\description{
Write priors for a linear mixed model
}
