% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_lme.R
\name{lmereg_priors}
\alias{lmereg_priors}
\title{Write priors for the regression coefficients of the linear mixed model}
\usage{
lmereg_priors(K, y_name)
}
\arguments{
\item{K}{K}

\item{y_name}{character string, name of outcome}
}
\description{
Write priors for the regression coefficients of the linear mixed model
}

