% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{match_positions}
\alias{match_positions}
\title{Find the position(s) of a variable in a model matrix}
\usage{
match_positions(varname, DF, colnams)
}
\arguments{
\item{varname}{character string to look for}

\item{DF}{data frame containing the variable}

\item{colnams}{column names of model matrix}
}
\value{
named vector specifying the columns in X that relate to varname
}
\description{
Find the position(s) of a variable in a model matrix
}
