% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_criteria.R
\name{MC_error}
\alias{MC_error}
\title{Monte Carlo error}
\usage{
MC_error(object, subset = "main", start = NULL, end = NULL, thin = NULL,
  ...)
}
\arguments{
\item{object}{JointAI object}

\item{...}{parameters passed to \code{\link[mcmcse]{mcmc.mat}}}
}
\description{
Monte Carlo error
}
