% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{merge_call_args}
\alias{merge_call_args}
\title{Merge call arguments with default formals}
\usage{
merge_call_args(formals, call, sframe)
}
\arguments{
\item{formals}{List of formal arguments for \verb{*_imp()}.}

\item{call}{The matched call from \verb{*_imp()} as returned by \code{match.call()}.}

\item{sframe}{The environment within \verb{*_imp()}
(obtained from \code{sys.frame(sys.nframe())})}
}
\value{
A list of arguments combining defaults and user-specified values.
}
\description{
Merge call arguments with default formals
}
\note{
Helper function for \link{prep_arglist}.
}
\keyword{internal}
