% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{normalize_formula_args}
\alias{normalize_formula_args}
\title{Normalize formula arguments in arglist}
\usage{
normalize_formula_args(arglist)
}
\arguments{
\item{arglist}{A list containing at least \code{formula}, \code{fixed}, and \code{random}
elements.}
}
\value{
The updated \code{arglist} with formulas converted to lists.
}
\description{
Normalize formula arguments in arglist
}
\note{
Helper function used in \code{\link[=prep_arglist]{prep_arglist()}}.
}
\keyword{internal}
