% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_divide_matrices.R
\name{paste_analysis_type}
\alias{paste_analysis_type}
\title{Paste analysis type with family information}
\usage{
paste_analysis_type(analysis_type)
}
\arguments{
\item{analysis_type}{character string, as created by the separate `*-imp()``
functions}
}
\value{
character string with family and link information appended
}
\description{
Paste analysis type with family information
}
\keyword{internal}
