% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGSmodel.R
\name{paste_coef}
\alias{paste_coef}
\title{Write the coefficient part of a linear predictor}
\usage{
paste_coef(parname, parelmts)
}
\arguments{
\item{parname}{character string; name of the coefficient (e.g., "beta")}

\item{parelmts}{vector of integers; the index of the parameter vector}
}
\value{
A vector of character strings of the form \code{beta[3]}.
}
\description{
Write the coefficient part of a linear predictor
}
\keyword{internal}
