% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGSmodel.R
\name{paste_data}
\alias{paste_data}
\title{Write the data element of a linear predictor}
\usage{
paste_data(matnam, index, col, isgk = FALSE)
}
\arguments{
\item{matnam}{characters string; name of the data matrix}

\item{index}{character string; the index (e.g., "i", or "ii")}

\item{col}{integer vector; the indices of the columns in \code{matnam}}

\item{isgk}{logical; is this for within the Gauss-Kronrod quadrature?}
}
\value{
A vector of character strings of the form
\code{M_id[i, 3]} or \code{M_id[i, 3, k]}.
}
\description{
Write the data element of a linear predictor
}
\keyword{internal}
