% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{paste_dummies}
\alias{paste_dummies}
\title{Write syntax to divide categorical variable into dummy variables in JAGS}
\usage{
paste_dummies(categories, dest_col, dummy_cols, ...)
}
\arguments{
\item{categories}{numeric vector of categories for which dummy variables
need to be created (i.e., excluding the reference category)}

\item{dest_col}{integer specifying the column in Xcat that contains the categorical variable}

\item{dummy_cols}{numeric vector specifying the columns in Xc that contain
the dummy variables corresponding to the categorical variable}
}
\description{
Write syntax to divide categorical variable into dummy variables in JAGS
}
