% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{paste_imp_model}
\alias{paste_imp_model}
\title{Call and paste imputation models}
\usage{
paste_imp_model(imp_par_list)
}
\arguments{
\item{imp_par_list}{list of parameters}
}
\description{
Call and paste imputation models
}
