% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{paste_imp_priors}
\alias{paste_imp_priors}
\title{Call and paste imputation priors}
\usage{
paste_imp_priors(imp_par_list)
}
\arguments{
\item{imp_par_list}{list of parameters}
}
\description{
Call and paste imputation priors
}
