% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{paste_interactions}
\alias{paste_interactions}
\title{Paste interaction terms for JAGS model}
\usage{
paste_interactions(index, mat0, mat1, mat0_col, mat1_col)
}
\description{
Paste interaction terms for JAGS model
}
