% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_impmodels.R
\name{paste_predictor}
\alias{paste_predictor}
\title{Build linear predictor for imputation models}
\usage{
paste_predictor(varname, par_els, Xc_cols, par_name, indent)
}
\arguments{
\item{varname}{character string, name of the variable to be imputed}

\item{par_els}{numeric vector specifying the elements of the vector of
regression parameters}

\item{Xc_cols}{numeric vector specifying the columns in the design matrix Xc}

\item{par_name}{character string, specifying the name of the regression
parameters (e.g. "alpha")}
}
\description{
Build linear predictor for imputation models
}

