% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGSmodel.R
\name{paste_scale}
\alias{paste_scale}
\title{Create the scaling in a data element of a linear predictor}
\usage{
paste_scale(x, row, scalemat)
}
\arguments{
\item{x}{a character string}

\item{row}{integer; indicating the row of \code{scalemat} to be used}

\item{scalemat}{character string; name of the matrix containing the scaling
information (e.g., "spM_lvlone"). This matrix is assumed to
have columns "center" and "scale".}
}
\value{
a character string of the form \code{(x - center)/scale}.
}
\description{
Create the scaling in a data element of a linear predictor
}
\keyword{internal}
