% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc_data.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Mayo Clinic Primary Biliary Cirrhosis, baseline data}
\format{A data frame with 418 rows and 19 variables:
\describe{
  \tabular{ll}{
  \code{age}      \tab in years\cr
  \code{albumin}  \tab serum albumin (g/dl)\cr
  \code{alk.phos} \tab alkaline phosphotase (U/liter)\cr
  \code{ascites}  \tab presence of ascites\cr
  \code{ast}      \tab aspartate aminotransferase, once called SGOT (U/ml)\cr
  \code{bili}     \tab serum bilirunbin (mg/dl)\cr
  \code{chol}     \tab	serum cholesterol (mg/dl)\cr
  \code{copper}   \tab urine copper (ug/day)\cr
  \code{edema}    \tab 0 no edema, 0.5 untreated or successfully treated,
                       1 edema despite diuretic therapy\cr
  \code{hepato}   \tab	presence of hepatomegaly or enlarged liver\cr
  \code{id}       \tab	case number\cr
  \code{platelet} \tab	platelet count\cr
  \code{protime}  \tab	standardised blood clotting time\cr
  \code{sex}      \tab	male (m) or female (f)\cr
  \code{spiders}  \tab	blood vessel malformations in the skin\cr
  \code{stage}    \tab	histologic stage of disease (needs biopsy)\cr
  \code{status}   \tab	status at endpoint, 0/1/2 for censored, transplant, dead\cr
  \code{time}     \tab	number of days between registration and the earlier of death,
                       transplantion, or study analysis in July, 1986\cr
  \code{trt}      \tab	1/2/NA for D-penicillmain, placebo, not randomised\cr
  \code{trig}     \tab triglycerides (mg/dl)\cr
  }
}}
\source{
Therneau, T. and Grambsch, P. (2000)
            \emph{Modeling Survival Data: Extending the Cox Model}.
            Springer-Verlag, New York.
}
\usage{
pbc
}
\description{
This data is from the Mayo Clinic trial in primary biliary cirrhosis (PBC) of the
liver conducted between 1974 and 1984.
The first 312 cases in the data set participated in the randomized placebo controlled
trial of the drug D-penicillamine and contain largely complete data.
The additional 106 cases did not participate in the clinical trial, but consented
to have basic measurements recorded and to be followed for survival
}
\examples{
summary(pbc)
}
\keyword{datasets}
