% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc_data.R
\docType{data}
\name{pbc_long}
\alias{pbc_long}
\title{Mayo Clinic Primary Biliary Cirrhosis, sequential data}
\format{A data frame with 1945 rows and 20 variables:
\describe{
  \tabular{ll}{
  \code{age}      \tab in years\cr
  \code{albumin}  \tab serum albumin (g/dl)\cr
  \code{alk.phos} \tab alkaline phosphotase (U/liter)\cr
  \code{ascites}  \tab presence of ascites\cr
  \code{ast}      \tab aspartate aminotransferase, once called SGOT (U/ml)\cr
  \code{bili}     \tab serum bilirunbin (mg/dl)\cr
  \code{chol}     \tab	serum cholesterol (mg/dl)\cr
  \code{day}      \tab number of days between enrollment and this visit date\cr
  \code{edema}    \tab 0 no edema, 0.5 untreated or successfully treated,
                       1 edema despite diuretic therapy\cr
  \code{futime}     \tab	number of days between registration and the earlier of death,
                       transplantion, or study analysis in July, 1986\cr
  \code{hepato}   \tab	presence of hepatomegaly or enlarged liver\cr
  \code{id}       \tab	case number\cr
  \code{platelet} \tab	platelet count\cr
  \code{protime}  \tab	standardised blood clotting time\cr
  \code{sex}      \tab	male (m) or female (f)\cr
  \code{spiders}  \tab	blood vessel malformations in the skin\cr
  \code{stage}    \tab	histologic stage of disease (needs biopsy)\cr
  \code{status}   \tab	status at endpoint, 0/1/2 for censored, transplant, dead\cr
  \code{trt}      \tab	1/2/NA for D-penicillmain, placebo, not randomised\cr
  }
}}
\source{
Therneau, T. and Grambsch, P. (2000)
            \emph{Modeling Survival Data: Extending the Cox Model}.
            Springer-Verlag, New York.
}
\usage{
pbc_long
}
\description{
This data is a continuation of the \link{pbc} data set, and contains the follow-up
laboratory data for the 312 randomized patients study patient.
Some baseline data values in this file differ from the original PBC file,
for instance, the data errors in prothrombin time and age which were discovered after
the orignal analysis (see Fleming and Harrington, figure 4.6.7).
}
\examples{
summary(pbc_long)
}
\keyword{datasets}
