% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.JointAI}
\alias{plot.JointAI}
\title{Plot an object object inheriting from class 'JointAI'}
\usage{
\method{plot}{JointAI}(x, ...)
}
\arguments{
\item{x}{object inheriting from class 'JointAI'}

\item{...}{currently not used}
}
\description{
Plot an object object inheriting from class 'JointAI'
}
\note{
Currently, \code{plot()} can only be used with (generalized) linear (mixed)
models.
}
\examples{
mod <- lm_imp(y ~ C1 + C2 + B1, data = wideDF, n.iter = 100)
plot(mod)

}
