% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{traceplot}
\alias{densplot}
\alias{traceplot.JointAI}
\alias{densplot.JointAI}
\title{Plot MCMC output from JointAI}
\usage{
traceplot(object, ...)

densplot(object, ...)

\method{traceplot}{JointAI}(object, start = NULL, end = NULL, thin = NULL,
  subset = "main", nrow = NULL, ncol = NULL, ...)

\method{densplot}{JointAI}(object, start = NULL, end = NULL, thin = NULL,
  subset = "main", vlines = NULL, nrow = NULL, ncol = NULL, ...)
}
\arguments{
\item{object}{a \code{JointAI}-object}

\item{start}{from which iteration onwards the MCMC samples should be included
(see \code{\link[coda]{window.mcmc}})}

\item{end}{last iteration included in the plot
(see \code{\link[coda]{window.mcmc}})}

\item{thin}{thinning interval (see \code{\link[coda]{window.mcmc}})}

\item{subset}{subset of monitored nodes (columns in the MCMC sample) to use.
Can be specified as a numeric vector of columns,  a vector of
column names, as \code{subset = "main"} or \code{NULL}.
If \code{NULL}, all monitored nodes will be plotted.
\code{subset = "main"} (default) the main parameters of the
analysis model will be plotted (regression coefficients,
standard deviation of the residual, random effects covariance
matrix).}

\item{nrow}{number of rows in the plotting layout; determined automatically
if not specified}

\item{ncol}{number of columns in the plotting layout; determined automatically
if not specified}

\item{vlines}{only for densplot; list where each element specifies parameters
to create a vertical line in each of the plots. The list
elements need to have a named element \code{v = [x location]}
and may have more named elements that can be passed to
\code{\link[graphics]{abline}}}
}
\description{
Creates a set of traceplots or densityplots from the MCMC sample of a JointAI-object
}
\examples{
\dontrun{
mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)
traceplot(mod1)
densplot(mod1)
densplot(mod1, vlines = list(list(v = rep(0, nrow(summary(mod1)$stats)),
                                  col = grey(0.8)),
                             list(v = summary(mod1)$stats[, "Mean"], lty = 1, lwd = 2),
                             list(v = summary(mod1)$stats[, "2.5\%"], lty = 2),
                             list(v = summary(mod1)$stats[, "97.5\%"], lty = 2))
        )
}

}
\seealso{
\code{\link{summary.JointAI}}, \code{\link{lme_imp}}, \code{\link{glm_imp}},
          \code{\link{lm_imp}}
}
