% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imp_distr.R
\name{plot_imp_distr}
\alias{plot_imp_distr}
\title{Plot the distribution of observed and imputed values}
\usage{
plot_imp_distr(
  data,
  imp = "Imputation_",
  id = ".id",
  rownr = ".rownr",
  ncol = NULL,
  nrow = NULL,
  labeller = NULL
)
}
\arguments{
\item{data}{a \code{data.frame} containing multiple imputations
and the original incomplete data stacked onto each other}

\item{imp}{the name of the variable specifying the imputation indicator}

\item{id}{the name of the variable specifying the subject indicator}

\item{rownr}{the name of a variable identifying which rows correspond to the
same observation in the original (unimputed) data}

\item{ncol}{Number of rows and columns.}

\item{nrow}{Number of rows and columns.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:label_parsed]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:label_value]{label_value()}} is used by default,
check it for more details and pointers to other options.}
}
\description{
Plots densities and barplots of the observed and imputed values in a
long-format dataset (multiple imputed datasets stacked onto each other).
}
\examples{
mod <- lme_imp(y ~ C1 + c2 + B2 + C2, random = ~ 1 | id, data = longDF,
               n.iter = 200, monitor_params = c(imps = TRUE), mess = FALSE)
impDF <- get_MIdat(mod, m = 5)
plot_imp_distr(impDF, id = "id", ncol = 3)

}
