% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ppc.R
\name{ppc_plot}
\alias{ppc_plot}
\title{Plot a posterior predictive check}
\usage{
ppc_plot(object, fun, vars = NULL, input = c("matrix", "scalar"))
}
\arguments{
\item{object}{object inheriting from class 'JointAI'}

\item{fun}{the function to be applied}

\item{vars}{optional; names of the variables for which the posterior predictive summary should be plotted}

\item{input}{type of input expected by the function; one of \code{"matrix"}
(iterations are rows, observations are columns) or \code{"scalar"}}
}
\description{
Plot a posterior predictive check
}
