% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ppc.R
\name{ppc_summary}
\alias{ppc_summary}
\title{Plot a posterior predictive summary}
\usage{
ppc_summary(object, fun, vars = NULL, type = c("histogram", "density",
  "chainwise_dens"), ...)
}
\arguments{
\item{object}{object inheriting from class 'JointAI'}

\item{fun}{the function to be used to summarize the sample}

\item{vars}{optional; names of the variables for which the posterior predictive summary should be plotted}

\item{type}{type of the plot; one of \code{"histogram"}, \code{"density"} or \code{"chainwise_density"} (may be abbreviated)}

\item{...}{additional parameters passed to the geom that is used}
}
\description{
Plot a posterior predictive summary
}
