% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{prep_arglist}
\alias{prep_arglist}
\title{Prepare list of arguments for model_imp()}
\usage{
prep_arglist(analysis_type, family = NULL, formals = formals(),
  call = match.call(), sframe = sys.frame(sys.nframe()))
}
\arguments{
\item{analysis_type}{Type of analysis to be performed.}

\item{family}{Family object or character string specifying the error distribution and link function.}

\item{formals}{List of formal arguments for the function.}

\item{call}{The matched call as returned by \code{match.call()}.}

\item{sframe}{An environment (typically from \code{sys.frame(sys.nframe())}) from which to retrieve argument values; usually the calling frame.}
}
\value{
A list of arguments prepared for model_imp(), including the analysis type, family, formulas, and other relevant parameters.
}
\description{
Prepare list of arguments for model_imp()
}
\keyword{internal}
