% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismodel_lme.R
\name{ranef_priors}
\alias{ranef_priors}
\title{Write priors for the random effects covariance matrix}
\usage{
ranef_priors(Z)
}
\arguments{
\item{Z}{design matrix of the random effects}
}
\description{
Write priors for the random effects covariance matrix
}

