% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_ranefs.R
\name{rd_terms_by_grouping}
\alias{rd_terms_by_grouping}
\title{Extract terms by grouping variables from a formula}
\usage{
rd_terms_by_grouping(formula)
}
\arguments{
\item{formula}{a \code{formula} object}
}
\value{
a named list of one-sided formulas, where each name corresponds to
a grouping variable and the formula contains all terms associated
with that grouping variable.
}
\description{
This function takes a formula as input, extracts terms that are
associated with grouping variables, and organizes them by grouping variable.
}
\keyword{internal}
