% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGS.R
\name{reformat_difftime}
\alias{reformat_difftime}
\title{Set all elements of a \code{difftime} object to the same, largest meaningful unit}
\usage{
reformat_difftime(dt)
}
\arguments{
\item{dt}{a \code{difftime} object (potentially a vector of \code{difftime}s)}
}
\description{
Set all elements of a \code{difftime} object to the same, largest meaningful unit
}
\keyword{internal}
