% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas.R
\name{remove_formula_grouping}
\alias{remove_formula_grouping}
\title{Remove grouping part from (random effects) formula}
\usage{
remove_formula_grouping(formula)
}
\arguments{
\item{formula}{A \code{formula} object (NOT a list of formulas)}
}
\value{
A list of one-sided \code{formula} objects without the grouping part,
split by grouping variable
}
\description{
Removes the part after (and including) the pipe symbol (\code{|}) in a formula.
}
\seealso{
\code{\link[=remove_grouping]{remove_grouping()}}
}
\keyword{internal}
