% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas.R
\name{remove_grouping}
\alias{remove_grouping}
\title{Remove grouping part from (random effects) formulas}
\usage{
remove_grouping(formulas)
}
\arguments{
\item{formulas}{A list of \code{formula} objects (or \code{NULL})}
}
\value{
A list of \code{formula} objects without the grouping part
}
\description{
Applies \code{remove_formula_grouping()} to a list of formulas.
}
\seealso{
\code{\link[=remove_formula_grouping]{remove_formula_grouping()}}
}
\keyword{internal}
