% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{remove_lhs}
\alias{remove_lhs}
\title{Remove the left hand side of a (list of) formula(s)}
\usage{
remove_lhs(formula)
}
\arguments{
\item{formula}{a formula object or a list of formula objects}
}
\value{
A \code{formula} object or a \code{list} of \code{formula} objects.
}
\description{
Internal function; used in divide_matrices, get_models and help functions
(2022-02-05)
}
\keyword{internal}
