% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{replace_nan_with_na}
\alias{replace_nan_with_na}
\title{Replace NaN values with NA}
\usage{
replace_nan_with_na(x)
}
\arguments{
\item{x}{a vector (also works for matrices and scalars)}
}
\value{
the vector (or object like the input) with \code{NaN} values replaced by
\code{NA}
}
\description{
Replace NaN values with NA
}
\keyword{internal}
