% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.JointAI}
\alias{residuals.JointAI}
\title{Extract residuals from an object of class JointAI}
\usage{
\method{residuals}{JointAI}(object, type = c("deviance", "response",
  "working"), ...)
}
\description{
Extract residuals from an object of class JointAI
}
