% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{resolve_family_obj}
\alias{resolve_family_obj}
\title{Resolve family object}
\usage{
resolve_family_obj(family)
}
\arguments{
\item{family}{Family object, character string, or function.}
}
\value{
A family object or NULL.
}
\description{
Converts a family specification (character string, function, or family
object) to a family object.
}
\note{
Helper function used in \code{\link[=prep_arglist]{prep_arglist()}}.
}
\keyword{internal}
