% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scaled_data.matrix}
\alias{scaled_data.matrix}
\title{calculate scaled data matrix}
\usage{
scaled_data.matrix(X, scale_vars, scale_pars)
}
\arguments{
\item{X}{a matrix}

\item{scale_vars}{a vector of variable names or FALSE}

\item{scale_pars}{a matrix of scaling parameters or FALSE or NULL}
}
\description{
calculate scaled data matrix
}
