% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{split_formula}
\alias{split_formula}
\title{Split a formula into fixed and random effects parts}
\usage{
split_formula(formula)
}
\arguments{
\item{formula}{a \code{formula} object}
}
\description{
Split a lme4 style formula into nlme style formulas.
}
\details{
Internal function, used in *_imp and help functions (2022-02-06)
}
\keyword{internal}
