% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_JAGS.R
\name{sum_duration}
\alias{sum_duration}
\title{Calculate the sum of the computational duration of a JointAI object}
\usage{
sum_duration(object, by = NULL)
}
\arguments{
\item{object}{object of class \code{JointAI}}

\item{by}{optional grouping information; options are \code{NULL} (default) to
calculate the sum over all chains and runs and both the adaptive
and sampling phase, \code{"run"} to get the duration per run,
\code{"phase"} to get the sum over all chains and runs per phase,
\code{"chain"} to get the sum per chain over both phases and all runs,
\code{"phase and run"} to get the sum over all chains, separately per
phase and run.}
}
\description{
Calculate the sum of the computational duration of a JointAI object
}
