% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions.R
\name{test_get_imp_method}
\alias{test_get_imp_method}
\title{## Tests for get_imp_method
Test that get_imp_method extracts all the incomplete variables and none of
the complete}
\usage{
test_get_imp_method(meth, fixed, random = NULL, DF, auxvars = NULL)
}
\description{
## Tests for get_imp_method
Test that get_imp_method extracts all the incomplete variables and none of
the complete
}

