% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{two_value_to_factor}
\alias{two_value_to_factor}
\title{Convert two-value vectors to factors}
\usage{
two_value_to_factor(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the vector converted to a factor if it has exactly two unique
(non-missing) values and is not already a factor; otherwise the input
}
\description{
Convert two-value vectors to factors
}
\keyword{internal}
