% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_divide_matrices.R
\name{varname_to_modelframe}
\alias{varname_to_modelframe}
\title{Create data.frame from variable term and data}
\usage{
varname_to_modelframe(varname, data)
}
\arguments{
\item{varname}{character string giving the variable name or model term}

\item{data}{a data.frame containing the variables mentioned in \code{varname}}
}
\value{
a \code{data.frame}
}
\description{
Create data.frame from variable term and data
}
\keyword{internal}
