% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wideDF}
\alias{wideDF}
\title{Cross-sectional example dataset}
\format{An object of class \code{data.frame} with 100 rows and 13 columns.}
\source{
simulated by \code{\link{simulate_data}}
}
\usage{
data(wideDF)
}
\description{
Cross-sectional example dataset
}
\keyword{datasets}
