% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_imp_spss.R
\name{write_imp_spss}
\alias{write_imp_spss}
\title{Write imputed dataset to a .sav file}
\usage{
write_imp_spss(object, m = 10, seed = NULL, resdir = NULL,
  filename = NULL, write_to_SPSS = T)
}
\arguments{
\item{object}{JointAI object}

\item{m}{number of imputed datasets}

\item{seed}{optional seed}

\item{resdir}{optional directory for results}

\item{filename}{optional file name (without ending)}

\item{write_to_SPSS}{logical}
}
\description{
Write imputed dataset to a .sav file
}
