% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Write model syntax to file}
\usage{
write_model(type, meth = NULL, Ntot, N, y_name, Mlist = NULL, Z = NULL,
  Xic = NULL, Xl = NULL, Xil = NULL, hc_list = NULL, K, imp_par_list,
  file = NULL, package = "JAGS")
}
\arguments{
\item{type}{analysis model type (character string)}

\item{meth}{named vector specifying imputation methods and ordering}

\item{Ntot}{number of observations}

\item{N}{number of individuals}

\item{y_name}{name of outcome variable}

\item{Mlist}{list of design matrices}

\item{Z}{random effects design matrix}

\item{Xic}{design matrix of time-constant interactions}

\item{Xl}{design matrix of time-varying covariates}

\item{Xil}{design matrix of interactions involving time-varying covariates}

\item{hc_list}{list specifying hierarchical centring structure}

\item{K}{matrix specifying range of regression coefficients used for each
component of the analysis model}

\item{file}{path and file name}

\item{package}{currently only JAGS is implemented}
}
\description{
Write model syntax to file
}

