% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{all_vars}
\alias{all_vars}
\title{Extract names of variables from a (list of) formula(s)}
\usage{
all_vars(fmla)
}
\arguments{
\item{fmla}{a formula or list of formulas}
}
\description{
Version of \code{all.vars()} that can handle lists of formulas.
}
\keyword{internal}
