% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas.R
\name{all_vars}
\alias{all_vars}
\title{Version of \code{all.vars()} that can handle lists of formulas}
\usage{
all_vars(fmla)
}
\arguments{
\item{a}{formula or list of formulas}
}
\description{
Version of \code{all.vars()} that can handle lists of formulas
}
\keyword{internal}
