% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{check_formula_list}
\alias{check_formula_list}
\title{Check/convert formula to list}
\usage{
check_formula_list(formula, convert = TRUE)
}
\arguments{
\item{formula}{any object}

\item{convert}{logical; should the input be converted to a list?}
}
\description{
Check if an object is a list of formulas and/or NULL elements and convert
it to a list if it is a formula object.
}
\details{
Internal function; used in many help functions, get_refs, *_imp, predict
(2022-02-05)
}
\keyword{internal}
