% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{densplot}
\alias{densplot}
\alias{densplot.JointAI}
\title{Densplots}
\usage{
densplot(x, ...)

\method{densplot}{JointAI}(x, start = NULL, end = NULL, thin = NULL,
  subset = NULL, vlines = NULL)
}
\arguments{
\item{x}{object}
}
\description{
Densplots
}
