% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{extract_id}
\alias{extract_id}
\title{Extract id variable from fixed effects formula}
\usage{
extract_id(fixed)
}
\arguments{
\item{fixed}{a formula describing fixed effects structure}
}
\value{
a character string, giving the name of the id variable or NULL
}
\description{
Extract id variable from fixed effects formula
}

