% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas_general.R
\name{extract_id}
\alias{extract_id}
\title{Extract all id variables from a list of random effects formulas}
\usage{
extract_id(random, warn = TRUE)
}
\arguments{
\item{random}{a one-sided random effects formula or a list of such formulas}

\item{warn}{logical; should warnings be printed?}
}
\description{
Internal function, used in \code{divide_matrices()}, \code{get_models()},
various help functions, \code{predict()} (2022-02-06)
}
\keyword{internal}
