% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_state.R
\name{extract_state}
\alias{extract_state}
\title{Return the current state of a 'JointAI' model}
\usage{
extract_state(object, pattern = paste0("^", c("RinvD", "invD", "tau", "b"),
  "_"))
}
\arguments{
\item{object}{an object of class 'JointAI'}

\item{pattern}{vector of patterns to be matched with the names of the nodes}
}
\description{
The current state of a preliminary model can be used as initial values
}
