% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{get_groups}
\alias{get_groups}
\title{Get grouping information}
\usage{
get_groups(idvars, data)
}
\arguments{
\item{idvars}{a character vector with the names of grouping variables}

\item{data}{a data.frame}
}
\value{
a list with grouping information for each grouping level:
each element is a vector of length \code{nrow(data)} with the group
membership indices of each observation for the corresponding
grouping level.
}
\description{
A helper function that generates grouping information from a data.frame
and a character vector with the names of grouping variables.
In all cases, the level "lvlone" is added to indicate the lowest level of
the data (i.e., each observation is its own group).
}
\keyword{internal}
