% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_vcov.R
\name{get_nranef}
\alias{get_nranef}
\title{Extract the number of random effects}
\usage{
get_nranef(random, data)
}
\arguments{
\item{random}{a random effect formula or list of random effects formulas}

\item{data}{a \code{data.frame}}
}
\value{
a list of named vectors of the numbers of random effects per variable,
were each list element refers to a (hierarchical) grouping level.
}
\description{
This function extracts the number of random effects per variable and grouping
level from the provided random effects formulas.
}
\keyword{internal}
