% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_criteria.R
\name{GR_crit}
\alias{GR_crit}
\title{Gelman-Rubin criterion for convergence}
\usage{
GR_crit(object, confidence = 0.95, transform = FALSE, autoburnin = TRUE,
  multivariate = TRUE, subset = NULL, start = NULL, end = NULL,
  thin = NULL, warn = TRUE, ...)
}
\arguments{
\item{object}{object inheriting from class \code{JointAI}}

\item{confidence}{the coverage probability of the confidence interval for the
potential scale reduction factor}

\item{transform}{a logical flag indicating whether variables in
\code{x} should be transformed to improve the normality of the
distribution. If set to TRUE, a log transform or logit transform, as
appropriate, will be applied.}

\item{autoburnin}{a logical flag indicating whether only the second half
  of the series should be used in the computation.  If set to TRUE
  (default) and \code{start(x)} is less than \code{end(x)/2} then start
  of series will be adjusted so that only second half of series is used.}

\item{multivariate}{a logical flag indicating whether the multivariate
potential scale reduction factor should be calculated for multivariate
chains}

\item{subset}{subset of parameters/variables/nodes (columns in the MCMC sample).
Uses the same logic as the argument \code{monitor_params} in
\code{\link{lm_imp}}, \code{\link{glm_imp}} and \code{\link{lme_imp}}.}

\item{start}{the first iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{end}{the last iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{thin}{thinning interval (see \code{\link[coda]{window.mcmc}})}

\item{warn}{logical; should warnings be given?; default is
\code{TRUE}. Note: this applies only to warnings
given directly by \strong{JointAI}.}

\item{...}{currently not used}
}
\description{
Gelman-Rubin criterion for convergence (uses \code{\link[coda]{gelman.diag}})
}
\examples{

mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)
GR_crit(mod1)

}
\references{
Gelman, A., Meng, X. L., & Stern, H. (1996).
Posterior predictive assessment of model fitness via realized discrepancies.
\emph{Statistica Sinica}, 733-760.
}
\seealso{
The vignette \href{https://nerler.github.io/JointAI/articles/SelectingParameters.html}{Selecting Parameters} contains some examples how to specify \code{subset}.
}
