% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAI.R
\docType{package}
\name{JointAI}
\alias{JointAI}
\alias{JointAI-package}
\title{JointAI: Joint Analysis and Imputation of Incomplete Data}
\description{
The \strong{JointAI} package performs simultaneous imputation and inference for
incomplete data using the Bayesian framework.
Distributions for incomplete variables are specified automatically and modeled
jointly with the analysis model.
}
\section{Main functions}{

The package has five main functions, \code{\link{lm_imp}},
\code{\link{glm_imp}}, \code{\link{lme_imp}}, \code{\link{glme_imp}} and \code{\link{survreg_imp}}
that allow analysis using linear regression, generalized linear regression,
linear mixed effects models, generalized linear mixed effects models and parametric
survival models. As far as possible, the specification of
the functions is the same as the specification of their complete data
versions \code{\link[stats]{lm}}, \code{\link[stats]{glm}},
\code{\link[nlme]{lme}} (from the package
\href{https://CRAN.R-project.org/package=nlme}{\strong{nlme}}) and
\code{\link[survival]{survreg}} (from the package
\href{https://CRAN.R-project.org/package=survival}{\strong{survival}}).


Results can be summarized and printed with \code{\link{summary.JointAI}},
and visualized using
\code{\link{traceplot}} or \code{\link{densplot}}.
The function \code{\link{predict.JointAI}} allows prediction (including credible intervals)
from \code{JointAI} models.
}

\section{Evaluation and export}{

Two criteria for evaluation of convergence and precision of the posterior
estimate are available:
\code{\link{GR_crit}} and \code{\link{MC_error}}.

Imputed data can be extracted and exported to SPSS data using \code{\link{get_MIdat}}.
}

\section{Other useful functions}{

\itemize{
\item \code{\link{parameters}} and \code{\link{list_impmodels}} to gain
      insight in the specified model
\item \code{\link{plot_all}} and \code{md_pattern} to visualize the
      distribution of the data and the missing data pattern
}
}

\section{Vignettes}{

The following vignettes are available
\itemize{
\item \href{https://nerler.github.io/JointAI/articles/MinimalExample.html}{\emph{Minimal Example}}:\cr
A minimal example demonstrating the use of
                           \code{\link{lm_imp}},
                           \code{\link{summary.JointAI}},
                           \code{\link{traceplot}}
                           and \code{\link{densplot}}.
\item \href{https://nerler.github.io/JointAI/articles/VisualizingIncompleteData.html}{\emph{Visualizing Incomplete Data}}:\cr
Demonstrations of the options in \code{\link{plot_all}} (plotting histograms
     and barplots for all variables in the data)  and \code{\link{md_pattern}}
     (plotting or printing the missing data pattern).

\item \href{https://nerler.github.io/JointAI/articles/ModelSpecification.html}{\emph{Model Specification}}:\cr
Explanation and demonstration of all parameters that are required or optional
to specify the model structure in \code{\link{lm_imp}}, \code{\link{glm_imp}}
and \code{\link{lme_imp}}.
Among others, the functions \code{\link{parameters}}, \code{\link{list_impmodels}},
\code{\link{get_imp_meth}} and \code{\link{set_refcat}} are used.

\item \href{https://nerler.github.io/JointAI/articles/SelectingParameters.html}{\emph{Parameter Selection}}:\cr
Examples on how to select the parameters/variables/nodes
                                to follow using the argument \code{monitor_params}
                                and the parameters/variables/nodes displayed
                                in the \code{\link{summary}}, \code{\link{traceplot}},
                                \code{\link{densplot}} or when using
                                \code{\link{GR_crit}} or \code{\link{MC_error}}.

\item \href{https://nerler.github.io/JointAI/articles/MCMCsettings.html}{\emph{MCMC Settings}}:\cr
Examples demonstrating how to set the arguments controlling settings of the MCMC sampling,
i.e., \code{n.adapt}, \code{n.iter}, \code{n.chains}, \code{thin}, \code{inits}.

\item \href{https://nerler.github.io/JointAI/articles/AfterFitting.html}{\emph{After Fitting}}:\cr
Examples on the use of functions to be applied after the model has been fitted,
including \code{\link{traceplot}}, \code{\link{densplot}}, \code{\link{summary}},
\code{\link{GR_crit}}, \code{\link{MC_error}}, \code{\link{predict}},
\code{\link{predDF}} and \code{\link{get_MIdat}}.
}
}

\references{
Erler, N. S., Rizopoulos, D., Rosmalen, J. V., Jaddoe,
V. W., Franco, O. H., & Lesaffre, E. M. (2016).
Dealing with missing covariates in epidemiologic studies: A comparison
between multiple imputation and a full Bayesian approach.
\emph{Statistics in Medicine}, 35(17), 2955-2974.
}
