% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{longDF}
\alias{longDF}
\title{Longitudinal example dataset}
\format{An object of class \code{data.frame} with 385 rows and 13 columns.}
\source{
simulated by \code{\link{simulate_data}}
}
\usage{
data(longDF)
}
\description{
Longitudinal example dataset
}
\keyword{datasets}
