% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_criteria.R
\name{MC_error}
\alias{MC_error}
\alias{plot.MCElist}
\title{Monte Carlo error}
\usage{
MC_error(
  x,
  subset = NULL,
  exclude_chains = NULL,
  start = NULL,
  end = NULL,
  thin = NULL,
  digits = 2,
  warn = TRUE,
  mess = TRUE,
  ...
)

\method{plot}{MCElist}(x, data_scale = TRUE, plotpars = NULL, ablinepars = list(v = 0.05), ...)
}
\arguments{
\item{x}{object inheriting from class 'JointAI'}

\item{subset}{subset of parameters/variables/nodes (columns in the MCMC sample).
Uses the same logic as the argument \code{monitor_params} in
\code{\link[JointAI:model_imp]{*_imp}}.}

\item{exclude_chains}{optional vector of the index numbers of chains that should be excluded}

\item{start}{the first iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{end}{the last iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{thin}{thinning interval (see \code{\link[coda]{window.mcmc}})}

\item{digits}{number of digits for output}

\item{warn}{logical; should warnings be given? Default is
\code{TRUE}. (Note: this applies only to warnings
given directly by \strong{JointAI}.)}

\item{mess}{logical; should messages be given? Default is
\code{TRUE}. (Note: this applies only to messages
given directly by \strong{JointAI}.)}

\item{...}{
  Arguments passed on to \code{\link[mcmcse:mcse.mat]{mcmcse::mcse.mat}}
  \describe{
    \item{\code{size}}{represents the batch size in ``bm'' and the truncation point in ``bartlett'' and ``tukey''. Default is \code{NULL} which implies that an optimal batch size is calculated using the \code{batchSize()} function. Can take character values of \code{``sqroot''} and \code{``cuberoot''} or any numeric value between 1 and n/2. \code{``sqroot''} means size is floor(n^(1/2)) and ``cuberoot'' means size is floor(n^(1/3)).}
    \item{\code{g}}{a function such that \eqn{E(g(x))} is the
  quantity of interest. The default is \code{NULL}, which
  causes the identity function to be used.}
    \item{\code{method}}{any of \code{``bm'', ``obm'', ``bartlett'', ``tukey''}. \code{``bm''} represents batch means estimator, \code{``obm''} represents overlapping batch means estimator with, \code{``bartlett''} and \code{``tukey''} represents the modified-Bartlett window and the Tukey-Hanning windows for  spectral variance estimators. }
    \item{\code{r}}{the lugsail parameter that converts a lag window into its lugsail equivalent. Larger values of \code{``r''} will typically imply less underestimation of \code{``cov''},  but higher variability of the estimator. Default is \code{``r = 3''} and \code{``r = 1,2''} are good choices. \code{``r > 5''} is not recommended. Non-integer values are ok.}
  }}

\item{data_scale}{show the Monte Carlo error of the sample transformed back
to the scale of the data (\code{TRUE}) or on the sampling scale (this
requires the argument \code{keep_scaled_mcmc = TRUE} in the JointAI model)}

\item{plotpars}{optional; list of parameters passed to \code{\link[graphics]{plot}()}}

\item{ablinepars}{optional; list of parameters passed to \code{\link[graphics]{abline}()}}
}
\value{
An object of class \code{MCElist} with elements \code{unscaled},
        \code{scaled} and \code{digits}. The first two are matrices with
        columns \code{est} (posterior mean), \code{MCSE} (Monte Carlo error),
        \code{SD} (posterior standard deviation) and \code{MCSE/SD}
        (Monte Carlo error divided by post. standard deviation.)
}
\description{
Calculate, print and plot the Monte Carlo error of the samples from a JointAI model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot Monte Carlo error
}}

\note{
Lesaffre & Lawson (2012) [p. 195] suggest the Monte Carlo error of a
      parameter should not be more than 5\% of the posterior standard
      deviation of this parameter (i.e., \eqn{MCSE/SD \le 0.05}).
}
\examples{
mod <- lm_imp(y ~ C1 + C2 + M2, data = wideDF, n.iter = 100)

MC_error(mod)

plot(MC_error(mod), ablinepars = list(lty = 2))

}
\references{
Lesaffre, E., & Lawson, A. B. (2012).
\emph{Bayesian Biostatistics}.
John Wiley & Sons.
}
\seealso{
The vignette \href{https://nerler.github.io/JointAI/articles/SelectingParameters.html}{Parameter Selection}
provides some examples how to specify the argument \code{subset}.
}
