% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_imp.R
\name{model_imp}
\alias{model_imp}
\alias{lm_imp}
\alias{lme_imp}
\alias{glm_imp}
\title{Joint analysis and imputation}
\usage{
lm_imp(fixed, data, auxvars = NULL, monitor_params = NULL,
  refcats = "largest", modelfile = NULL, overwrite = F, n.chains = 3,
  n.adapt = 100, n.iter = 0, thin = 1, runMCMC = F,
  MCMCpackage = "JAGS", meth = NULL, Mlist = NULL, K = NULL,
  K_imp = NULL, imp_pos = NULL, dest_cols = NULL, imp_par_list = NULL,
  scale_vars = NULL, scale_pars = NULL)

lme_imp(fixed, data, random, auxvars = NULL, monitor_params = NULL,
  refcats = "largest", modelfile = NULL, overwrite = F, n.chains = 3,
  n.adapt = 100, n.iter = 0, thin = 1, runMCMC = F,
  MCMCpackage = "JAGS", meth = NULL, Mlist = NULL, K = NULL,
  K_imp = NULL, imp_pos = NULL, dest_cols = NULL, imp_par_list = NULL,
  scale_vars = NULL, scale_pars = NULL)

glm_imp(fixed, family, data, auxvars = NULL, monitor_params = NULL,
  refcats = "largest", modelfile = NULL, overwrite = F, n.chains = 3,
  n.adapt = 100, n.iter = 0, thin = 1, runMCMC = F,
  MCMCpackage = "JAGS", meth = NULL, Mlist = NULL, K = NULL,
  K_imp = NULL, imp_pos = NULL, dest_cols = NULL, imp_par_list = NULL,
  scale_vars = NULL, scale_pars = NULL)
}
\arguments{
\item{fixed}{a two sided formula describing the fixed-effects part of the
model (see \code{\link[stats]{formula}})}

\item{data}{a data frame containing the variabels named in \code{fixed}}

\item{auxvars}{an optional vector of variable names that shoud be included
as predictors in the imputation procedure but are not part of
the analysis model}

\item{monitor_params}{a character vector giving the names of variables to be monitored, see details}

\item{refcats}{a named vector specifying which category should be treated as
as reference category for each categorical variable, "first"
or "largest" (default). When \code{refcat = "largest"}, the
category with the most observations is chosen for each
categorical variable.}

\item{modelfile}{optional name (and path) of the file the JAGS model will be
written to}

\item{overwrite}{logical}

\item{n.chains}{the number of parallel chains for the model}

\item{n.adapt}{the number of iterations for adaptation. See 
    \code{\link{adapt}} for details. If \code{n.adapt = 0} then no 
    adaptation takes place.}

\item{n.iter}{number of iterations to monitor}

\item{thin}{thinning interval for monitors}

\item{runMCMC}{logical}

\item{scale_vars}{NULL or a vector of variables to scale}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link{family}} for details of
    family functions.)}
}
\description{
Joint analysis and imputation
}
