% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_imp.R
\name{model_imp}
\alias{model_imp}
\alias{lm_imp}
\alias{glm_imp}
\alias{lme_imp}
\title{Joint analysis and imputation of incomplete data}
\usage{
lm_imp(formula, data, n.chains = 3, n.adapt = 100, n.iter = 0,
  thin = 1, monitor_params = NULL, inits = TRUE, modelname = NULL,
  modeldir = NULL, overwrite = NULL, keep_model = FALSE,
  quiet = TRUE, progress.bar = "text", warn = TRUE, mess = TRUE,
  auxvars = NULL, meth = NULL, refcats = NULL, trunc = NULL,
  scale_vars = NULL, hyperpars = NULL, ...)

glm_imp(formula, family, data, n.chains = 3, n.adapt = 100,
  n.iter = 0, thin = 1, monitor_params = NULL, inits = TRUE,
  modelname = NULL, modeldir = NULL, overwrite = NULL,
  keep_model = FALSE, quiet = TRUE, progress.bar = "text",
  warn = TRUE, mess = TRUE, auxvars = NULL, meth = NULL,
  refcats = NULL, trunc = NULL, scale_vars = NULL,
  hyperpars = NULL, ...)

lme_imp(fixed, data, random, n.chains = 3, n.adapt = 100, n.iter = 0,
  thin = 1, monitor_params = NULL, inits = TRUE, modelname = NULL,
  modeldir = NULL, overwrite = NULL, keep_model = FALSE,
  quiet = TRUE, progress.bar = "text", warn = TRUE, mess = TRUE,
  auxvars = NULL, meth = NULL, refcats = NULL, trunc = NULL,
  scale_vars = NULL, hyperpars = NULL, ...)
}
\arguments{
\item{formula}{a two sided model formula (see \code{\link[stats]{formula}})}

\item{data}{a data frame}

\item{n.chains}{the number of parallel chains for the model}

\item{n.adapt}{the number of iterations for adaptation. See 
    \code{\link{adapt}} for details. If \code{n.adapt = 0} then no 
    adaptation takes place.}

\item{n.iter}{number of iterations to monitor}

\item{thin}{thinning interval for monitors}

\item{monitor_params}{named vector specifying which parameters should be
monitored, see details.}

\item{inits}{optional specification of initial values in the form of a list
or a function (see \code{\link[rjags]{jags.model}}.
If omitted, initial values will be generated automatically.
It is an error to supply an initial value for an observed node.}

\item{modelname}{optional; character string specifying the name of model file
(including the ending, either .R or .txt).
If unspecified a random name will be generated.}

\item{modeldir}{optional; directory containing model file. If unspecified a
temporary directory will be created.}

\item{overwrite}{logical; whether an existing model file with the specified
\code{modeldir/modelname} should be overwritten. If set to
\code{FALSE} (default) and a model already exists, that
model will be used.}

\item{keep_model}{logical; whether the created JAGS model should be saved
or removed from the disk (\code{FALSE}; default) when the
sampling has finished.}

\item{quiet}{if \code{TRUE} then messages generated during compilation
    will be suppressed, as well as the progress bar during adaptation.}

\item{progress.bar}{type of progress bar. Possible values are
    \code{"text"}, \code{"gui"}, and \code{"none"}. See Details.}

\item{warn}{logical; should warnings be given?; default is
\code{TRUE}. Note: this applies only to warnings
given directly by \strong{JointAI}.}

\item{mess}{logical; should messages be given?; default is
\code{TRUE}. Note: this applies only to messages
given directly by \strong{JointAI}.}

\item{auxvars}{optional vector of variable names that should be used as
predictors in the imputation procedure (and will be imputed
if necessary) but are not part of the analysis model}

\item{meth}{optional named vector specifying imputation model types and order.
If \code{NULL} (default) imputation models will be determined
automatically based on the class of the columns of \code{data}
that contain missing values (see Details).
The default order is according to the proportion of missing
values (increasing).}

\item{refcats}{optional; a named list specifying which category should be
used as reference category for each of the categorical variables.
Options are the category label, the category number,
'first" (the first category) or "largest" (chooses the
category with the most observations). Default is "first".}

\item{trunc}{optional named list specifying the limits of truncation for the
distribution of the named incomplete variables
of each}

\item{scale_vars}{optional; named vector of (continuous) variables that will
be scaled (such that mean = 0 and sd = 1) to improve
convergence of the MCMC sampling. Default is that all
continuous variables that are not transformed by a function
(e.g. \code{log(), ns()}) will be scaled. Variables
for which \code{"lognorm"} is set as imputation model are
only scaled with regards to the standard deviation, but not
centered. If set to \code{FALSE} no scaling will be done.}

\item{hyperpars}{list of hyperparameters, as obtained by \code{\link{default_hyperpars}()};
only needs to be supplied if hyperparameters other than the
default should be used}

\item{...}{additional, optional arguments, see below}

\item{family}{only for \code{glm_imp}:
a description of the distribution and link function to
be used in the model. This can be a character string naming a
family function, a family function or the result of a call to
a family function. (See \code{\link[stats]{family}} and the
`Details` section below.)}

\item{fixed}{a two sided formula describing the fixed-effects part of the
model (see \code{\link[stats]{formula}})}

\item{random}{only for \code{lme_imp}:
a one-sided formula of the form \code{~x1 + ... + xn | g},
where \code{x1 + ... + xn} specifies the model for the random
effects and \code{g} the grouping variable}
}
\value{
An object of class \code{JointAI}
}
\description{
\code{lm_imp}, \code{glm_imp} and \code{lme_imp} estimate linear, generalized
linear and linear mixed models, respectively, using MCMC sampling.
}
\section{Optional arguments}{

There are some optional parameters that can be passed to \code{\dots}
\describe{
\item{\code{scale_pars}}{optional matrix of parameters used for centering and
scaling continuous covariates. If not specified, this will be calculated
automatically. If \code{FALSE}, no scaling will be done.}
}
}

\section{Details}{

\subsection{Implemented distribution families and link functions for \code{glm_imp()}}{
\tabular{ll}{
\code{gaussian} \tab with links: \code{identity}, \code{log}\cr
\code{binomial} \tab with links: \code{logit}, \code{probit}, \code{log}, \code{cloglog}\cr
\code{Gamma}    \tab with links: \code{identity}, \code{log}\cr
\code{poisson}  \tab with links: \code{log}, \code{identity}
}}



\subsection{Imputation methods}{
Implemented imputation models that can be chosen in the argument \code{meth} are:
\tabular{ll}{
\code{norm} \tab linear model\cr
\code{lognorm} \tab log-linear model for skewed continuous data\cr
\code{gamma} \tab gamma model (with log-link) for skewed continuous data\cr
\code{beta} \tab beta model (with logit-link) for skewed continuous data in (0, 1)\cr
\code{logit} \tab logistic model for binary data\cr
\code{multilogit} \tab multinomial logit model for unordered categorical variables\cr
\code{cumlogit} \tab cumulative logit model for ordered categorical variables
}}

\subsection{Parameters to follow (\code{monitor_params})}{
See also the vignette: \href{https://nerler.github.io/JointAI/articles/SelectingParameters.html}{Selecting Parameters}

Named vector specifying which parameters should be monitored. This can be done
either directly by specifying the name of the parameter or indirectly by one
of the key words summarizing a number of parameters. Except for \code{other},
in which parameter names are specified directly, parameter (groups) are just
set as \code{TRUE} or \code{FALSE}.
If left unspecified, \code{monitor_params = c("analysis_main" = TRUE)} will be used.
\tabular{ll}{
\strong{name/key word} \tab \strong{what is monitored}\cr
\code{analysis_main} \tab \code{betas}, \code{tau_y} and \code{sigma_y}\cr
\code{analysis_random} \tab \code{ranef}, \code{D}, \code{invD}, \code{RinvD}\cr
\code{imp_pars} \tab \code{alphas}, \code{tau_imp}, \code{gamma_imp}, \code{delta_imp}\cr
\code{imps} \tab imputed values\cr
\code{betas} \tab regression coefficients of the analysis model\cr
\code{tau_y} \tab precision of the residuals from the analysis model\cr
\code{sigma_y} \tab standard deviation of the residuals from the analysis model\cr
\code{ranef} \tab random effects\cr
\code{D} \tab covariance matrix of the random effects\cr
\code{invD} \tab inverse of \code{D}\cr
\code{RinvD} \tab matrix in prior for \code{invD}\cr
\code{alphas} \tab regression coefficients in the imputation models\cr
\code{tau_imp} \tab precision parameters of the residuals from imputation models\cr
\code{gamma_imp} \tab intercepts in ordinal imputation models\cr
\code{delta_imp} \tab increments of ordinal intercepts\cr
\code{other} \tab additional parameters
}
For example:

\code{monitor_params = c("analysis_main" = TRUE, "tau_y" = FALSE)}
would monitor the regression parameters \code{betas} and residual standard
deviation \code{sigma_y}, but not the residual precision.

\code{monitor_params = c(imps = TRUE)} would monitor \code{betas}, \code{tau_y},
and \code{sigma_y} (because \code{analysis_main = TRUE} by default) as well as
the imputed values.
}
}

\examples{

mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)
mod2 <- glm_imp(B1 ~ C1 + C2 + M2, data = wideDF,
                family = binomial(link = "logit"), n.iter = 100)
mod3 <- lme_imp(y ~ C1 + B2 + L1 + time, random = ~ time|id,
                data = longDF, n.iter = 500)


}
\seealso{
\code{\link{set_refcat}},
\code{\link{traceplot}}, \code{\link{densplot}},
         \code{\link{summary.JointAI}}, \code{\link{MC_error}},
         \code{\link{GR_crit}}, \code{\link[rjags]{jags.model}},
         \code{\link[rjags]{coda.samples}}, \code{predict.JointAI}
}
