% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc_data.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Mayo Clinic Primary Biliary Cirrhosis, baseline data}
\format{A data frame with 418 rows and 19 variables:
\describe{
  \item{age}{in years}
  \item{albumin}{serum albumin (g/dl)}
  \item{alk.phos}{alkaline phosphotase (U/liter)}
  \item{ascites}{presence of ascites}
  \item{ast}{aspartate aminotransferase, once called SGOT (U/ml)}
  \item{bili}{serum bilirubin (mg/dl)}
  \item{chol}{serum cholesterol (mg/dl)}
  \item{copper}{urine copper (ug/day)}
  \item{edema}{0 no edema, 0.5 untreated or successfully treated,
                       1 edema despite diuretic therapy}
  \item{hepato}{presence of hepatomegaly or enlarged liver}
  \item{id}{case number}
  \item{platelet}{platelet count}
  \item{protime}{standardized blood clotting time}
  \item{sex}{male (m) or female (f)}
  \item{spiders}{blood vessel malformations in the skin}
  \item{stage}{histologic stage of disease (needs biopsy)}
  \item{status}{status at endpoint, 0/1/2 for censored, transplant, dead}
  \item{time}{number of days between registration and the earlier of death,
                       transplantation, or study analysis in July, 1986}
  \item{trt}{1/2/NA for D-penicillamine, placebo, not randomized}
  \item{trig}{triglycerides (mg/dl)}
  }}
\source{
Therneau, T. and Grambsch, P. (2000)
            \emph{Modeling Survival Data: Extending the Cox Model}.
            Springer-Verlag, New York.
}
\usage{
pbc
}
\description{
This data is from the Mayo Clinic trial in primary biliary cirrhosis (PBC) of the
liver conducted between 1974 and 1984.
The first 312 cases in the data set participated in the randomized placebo controlled
trial of the drug D-penicillamine and contain largely complete data.
The additional 106 cases did not participate in the clinical trial, but consented
to have basic measurements recorded and to be followed for survival
}
\examples{
summary(pbc)
}
\keyword{datasets}
\keyword{internal}
