% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc_data.R
\docType{data}
\name{pbc_long}
\alias{pbc_long}
\title{Mayo Clinic Primary Biliary Cirrhosis, sequential data}
\format{A data frame with 1945 rows and 20 variables:
\describe{
  \item{age}{in years}
  \item{albumin}{serum albumin (g/dl)}
  \item{alk.phos}{alkaline phosphotase (U/liter)}
  \item{ascites}{presence of ascites}
  \item{ast}{aspartate aminotransferase, once called SGOT (U/ml)}
  \item{bili}{serum bilirubin (mg/dl)}
  \item{chol}{serum cholesterol (mg/dl)}
  \item{day}{number of days between enrollment and this visit date}
  \item{edema}{0 no edema, 0.5 untreated or successfully treated,
                       1 edema despite diuretic therapy}
  \item{futime}{number of days between registration and the earlier of death,
                       transplantation, or study analysis in July, 1986}
  \item{hepato}{presence of hepatomegaly or enlarged liver}
  \item{id}{case number}
  \item{platelet}{platelet count}
  \item{protime}{standardized blood clotting time}
  \item{sex}{male (m) or female (f)}
  \item{spiders}{blood vessel malformations in the skin}
  \item{stage}{histologic stage of disease (needs biopsy)}
  \item{status}{status at endpoint, 0/1/2 for censored, transplant, dead}
  \item{trt}{1/2/NA for D-penicillamine, placebo, not randomized}
}}
\source{
Therneau, T. and Grambsch, P. (2000)
            \emph{Modeling Survival Data: Extending the Cox Model}.
            Springer-Verlag, New York.
}
\usage{
pbc_long
}
\description{
This data is a continuation of the \link{pbc} data set, and contains the follow-up
laboratory data for the 312 randomized patients study patient.
Some baseline data values in this file differ from the original PBC file,
for instance, the data errors in prothrombin time and age which were discovered after
the original analysis (see Fleming and Harrington, figure 4.6.7).
}
\examples{
summary(pbc_long)
}
\keyword{datasets}
\keyword{internal}
