% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imp_distr.R
\name{plot_imp_distr}
\alias{plot_imp_distr}
\title{Plot the distribution of observed and imputed values}
\usage{
plot_imp_distr(data, imp = "Imputation_", id = "id", ncol = NULL,
  nrow = NULL)
}
\arguments{
\item{data}{a data.frame containing multiple imputations}

\item{imp}{the name of the variable specifying the imputation indicator}

\item{id}{the name of the variable specifying the subject indicator}

\item{ncol}{optional number of rows and columns in the plot layout;
automatically chosen if unspecified}

\item{nrow}{optional number of rows and columns in the plot layout;
automatically chosen if unspecified}
}
\description{
Plot the distribution of observed and imputed values
}
