% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{remove_grouping}
\alias{remove_grouping}
\title{Remove grouping from formula}
\usage{
remove_grouping(fmla)
}
\arguments{
\item{fmla}{a formula object or a string describing a formula}
}
\value{
returns the formula without the grouping part, i.e., with out
        anything behind "|"
}
\description{
Remove grouping from formula
}
