% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_formulas.R
\name{remove_lhs}
\alias{remove_lhs}
\title{Remove the left hand side of a (list of) formula(s)}
\usage{
remove_lhs(formula)
}
\arguments{
\item{formula}{a formula object or a list of formula objects}
}
\description{
Remove the left hand side of a (list of) formula(s)
}
\keyword{internal}
