% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_checks.R
\name{resolve_family_obj}
\alias{resolve_family_obj}
\title{Resolve family object}
\usage{
resolve_family_obj(family)
}
\arguments{
\item{family}{Family object, character string, or function.}
}
\value{
A family object or NULL.
}
\description{
Resolve family object
}
\keyword{internal}
