% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAI.R
\name{sharedParams}
\alias{sharedParams}
\title{Parameters used by several functions in JointAI}
\arguments{
\item{object}{object inheriting from class \code{JointAI}}

\item{subset}{subset of parameters/variables/nodes (columns in the MCMC sample).
Uses the same logic as the argument \code{monitor_params} in
\code{\link{lm_imp}}, \code{\link{glm_imp}} and \code{\link{lme_imp}}.}

\item{start}{the first iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{end}{the last iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{thin}{thinning interval (see \code{\link[coda]{window.mcmc}})}

\item{nrow, ncol}{optional number of rows and columns in the plot layout;
automatically chosen if unspecified}

\item{use_ggplot}{logical; Should ggplot be used instead of the base graphics?}

\item{warn}{logical; should warnings be given?; default is
\code{TRUE}. Note: this applies only to warnings
given directly by \strong{JointAI}.}

\item{mess}{logical; should messages be given?; default is
\code{TRUE}. Note: this applies only to messages
given directly by \strong{JointAI}.}

\item{xlab, ylab}{labels for the x- and y-axis}

\item{use_levels}{logical; should the multi-level structure be taken into account?
requires specification of the argument \code{idvar}}

\item{idvar}{name of the column that specifies the multi-level grouping structure}
}
\description{
Parameters used by several functions in JointAI
}
