% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.JointAI.R
\name{summary.JointAI}
\alias{summary.JointAI}
\alias{print.summary.JointAI}
\title{Summary of an object of class JointAI}
\usage{
\method{summary}{JointAI}(object, start = NULL, end = NULL, thin = NULL,
  quantiles = c(0.025, 0.975), subset = NULL, warn = TRUE, mess = TRUE,
  ...)

\method{print}{summary.JointAI}(x, digits = max(3, .Options$digits - 3), ...)
}
\arguments{
\item{object}{object inheriting from class \code{JointAI}}

\item{start}{the first iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{end}{the last iteration of interest (see \code{\link[coda]{window.mcmc}})}

\item{thin}{thinning interval (see \code{\link[coda]{window.mcmc}})}

\item{quantiles}{posterior quantiles}

\item{subset}{subset of parameters/variables/nodes (columns in the MCMC sample).
Uses the same logic as the argument \code{monitor_params} in
Defauls to the parameters of the analysis model.}

\item{warn}{logical; should warnings be given?; default is
\code{TRUE}. Note: this applies only to warnings
given directly by \strong{JointAI}.}

\item{mess}{logical; should messages be given?; default is
\code{TRUE}. Note: this applies only to messages
given directly by \strong{JointAI}.}

\item{\dots}{currently not used}

\item{x}{an object of class \code{summary.JointAI}}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link{print.default}}.}
}
\description{
Summary of an object of class JointAI
}
\examples{
mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)
summary(mod1)


}
\seealso{
The model fitting functions \code{\link{lm_imp}},
         \code{\link{glm_imp}}, \code{\link{lme_imp}}
}
