% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.JointAI}
\title{Traceplots}
\usage{
traceplot(x, ...)

\method{traceplot}{JointAI}(x, start = NULL, end = NULL, thin = NULL,
  subset = NULL)
}
\arguments{
\item{x}{object}
}
\description{
Traceplots
}
