% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.JointAI}
\title{Traceplots}
\usage{
traceplot(object, ...)

\method{traceplot}{JointAI}(object, start = NULL, end = NULL, thin = NULL,
  subset = "main", nrow = NULL, ncol = NULL, ...)
}
\arguments{
\item{object}{object}
}
\description{
Traceplots
}
