% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wideDF}
\alias{wideDF}
\title{Cross-sectional example dataset}
\format{A simulated data frame with 100 rows and 13 variables:
\describe{
  \item{C1}{continuous, complete variable}
  \item{C2}{continuous, incomplete variable}
  \item{B1}{binary, complete variable}
  \item{B2}{binary, incomplete variable}
  \item{M1}{unordered factor; complete variable}
  \item{M2}{unordered factor; incomplete variable}
  \item{O1}{ordered factor; complete variable}
  \item{O2}{ordered factor; incomplete variable}
  \item{L1}{continuous, complete variable}
  \item{L2}{continuous incomplete variable}
  \item{id}{id (grouping) variable}
  \item{time}{continuous complete variable}
  \item{y}{continuous, complete variable}
}}
\usage{
data(wideDF)
}
\description{
A simulated cross-sectional dataset.
}
\keyword{datasets}
